Add --animate option to subsim-server
This option isn't perfect.  It shows the animation after the fact, so
subs that are shot at already show the damage of the shot (as the shot
is being animated) and subs that are destroyed by a torpedo shot are
already removed from the map by the time the animation is drawn.

This option is mostly for amusement purposes, so at present I have no
intentions to fix the problem desribed above.